requirements.txt
Flask==2.1.2
Flask-Passlib==0.1
Flask-WTF==1.0.1
Flask-WkHTMLtoPDF==0.1.0
Jinja2==3.1.2
PyMySQL==1.0.2
SQLAlchemy==1.4.37
configparser==5.2.0
crypto==1.4.1
fastapi==0.78.0
getmac==0.8.3
numpy==1.22.4
pycryptodome==3.14.1
ecdsa==0.17.0
pyinput==0.3.2
pyasn1==0.4.8
rsa==4.8
python-jose==3.3.0
uvicorn==0.17.6